hirescope-streamlit/requirements.txt
streamlit
google.generativeai
python-dotenv
speechrecognition
pyaudio